backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.0.0
pyttsx3==2.90
Pillow==10.0.0